+++ app/api/dao/user.py
@@ -14,9 +14,8 @@ from app.utils.decorator_utils import email_verification_required
 from app.utils.enum_utils import MentorshipRelationState
 from app.database.models.mentorship_relation import MentorshipRelationModel
 from app.api.models.mentorship_relation import (
-    list_tasks_response_body,
-    mentorship_request_response_body_for_user_dashboard_body,
-)
+        mentorship_request_response_body_for_user_dashboard_body)
+from app.api.models.task import (list_tasks_response_body)
 from app.api.dao.mentorship_relation import MentorshipRelationDAO
 from app.utils.validation_utils import is_email_valid
