+++ app/api/models/mentorship_relation.py
@@ -11,16 +11,11 @@ def add_models_to_namespace(api_namespace):
         mentorship_request_response_body.name
     ] = mentorship_request_response_body
     api_namespace.models[relation_user_response_body.name] = relation_user_response_body
-    api_namespace.models[create_task_request_body.name] = create_task_request_body
-    api_namespace.models[list_tasks_response_body.name] = list_tasks_response_body
     api_namespace.models[
         mentorship_request_response_body_for_user_dashboard_body.name
     ] = mentorship_request_response_body_for_user_dashboard_body
     api_namespace.models[user_dashboard_user_details.name] = user_dashboard_user_details
-    api_namespace.models[task_comment_model.name] = task_comment_model
-    api_namespace.models[task_comments_model.name] = task_comments_model
-
-
+   
 send_mentorship_request_body = Model(
     "Send mentorship relation request model",
     {
@@ -84,34 +79,6 @@ mentorship_request_response_body = Model(
     },
 )
 
-create_task_request_body = Model(
-    "Create task request model",
-    {
-        "description": fields.String(
-            required=True, description="Mentorship relation task description"
-        )
-    },
-)
-
-list_tasks_response_body = Model(
-    "List tasks response model",
-    {
-        "id": fields.Integer(required=True, description="Task ID"),
-        "description": fields.String(
-            required=True, description="Mentorship relation task description"
-        ),
-        "is_done": fields.Boolean(
-            required=True, description="Mentorship relation task is done indication"
-        ),
-        "created_at": fields.Float(
-            required=True, description="Task creation date in UNIX timestamp format"
-        ),
-        "completed_at": fields.Float(
-            required=False, description="Task completion date in UNIX timestamp format"
-        ),
-    },
-)
-
 user_dashboard_user_details = Model(
     "user details for dashboard",
     {
@@ -156,27 +123,5 @@ mentorship_request_response_body_for_user_dashboard_body = Model(
             description="Mentorship relation state",
         ),
         "notes": fields.String(required=True, description="Mentorship relation notes"),
-    },
-)
-
-task_comment_model = Model(
-    "Task comment model.",
-    {"comment": fields.String(required=True, description="Task comment.")},
-)
-
-task_comments_model = Model(
-    "Task comments model.",
-    {
-        "id": fields.Integer(required=True, description="Task comment's id."),
-        "user_id": fields.Integer(required=True, description="User's id."),
-        "task_id": fields.Integer(required=True, description="Task's id."),
-        "relation_id": fields.Integer(required=True, description="Relation's id."),
-        "creation_date": fields.Float(
-            required=True, description="Creation date of the task comment."
-        ),
-        "modification_date": fields.Float(
-            required=False, description="Modification date of the task comment."
-        ),
-        "comment": fields.String(required=True, description="Task comment."),
     },
 )
