+++ app/api/resources/mentorship_relation.py
@@ -495,353 +495,3 @@ class ListPendingMentorshipRequests(Resource):
         return response
 
 
-@mentorship_relation_ns.route("mentorship_relation/<int:request_id>/task")
-class CreateTask(Resource):
-    @classmethod
-    @jwt_required
-    @mentorship_relation_ns.doc("create_task_in_mentorship_relation")
-    @mentorship_relation_ns.expect(auth_header_parser, create_task_request_body)
-    @mentorship_relation_ns.response(HTTPStatus.CREATED, '%s'%messages.TASK_WAS_CREATED_SUCCESSFULLY)
-    @mentorship_relation_ns.response(HTTPStatus.FORBIDDEN, '%s'%messages.UNACCEPTED_STATE_RELATION)
-    @mentorship_relation_ns.response(HTTPStatus.UNAUTHORIZED, '%s\n%s\n%s'%(
-        messages.TOKEN_HAS_EXPIRED,
-        messages.TOKEN_IS_INVALID,
-        messages.AUTHORISATION_TOKEN_IS_MISSING
-        )
-    )
-    @mentorship_relation_ns.response(403, '%s'%messages.USER_NOT_INVOLVED_IN_THIS_MENTOR_RELATION)
-    def post(cls, request_id):
-        """
-        Create a task for a mentorship relation.
-
-        Input:
-        1. Header: valid access token
-        2. Path: ID of request for which task is being created (request_id)
-        3. Body: JSON object containing description of task.
-
-        Returns:
-        Success or failure message. It gets added to GET /tasks endpoint and
-        is visible to the other person in the mentorship relation.
-        """
-
-        # TODO check if user id is well parsed, if it is an integer
-
-        user_id = get_jwt_identity()
-        request_body = request.json
-
-        is_valid = CreateTask.is_valid_data(request_body)
-
-        if is_valid != {}:
-            return is_valid, HTTPStatus.BAD_REQUEST
-
-        response = TaskDAO.create_task(
-            user_id=user_id, mentorship_relation_id=request_id, data=request_body
-        )
-
-        return response
-
-    @staticmethod
-    def is_valid_data(data):
-
-        if "description" not in data:
-            return messages.DESCRIPTION_FIELD_IS_MISSING
-
-        return {}
-
-
-@mentorship_relation_ns.route("mentorship_relation/<int:request_id>/task/<int:task_id>")
-class DeleteTask(Resource):
-    @classmethod
-    @jwt_required
-    @mentorship_relation_ns.doc("delete_task_in_mentorship_relation")
-    @mentorship_relation_ns.expect(auth_header_parser)
-    @mentorship_relation_ns.response(HTTPStatus.OK, "%s" % messages.TASK_WAS_DELETED_SUCCESSFULLY)
-    @mentorship_relation_ns.response(
-        HTTPStatus.UNAUTHORIZED,
-        "%s\n%s\n%s\n%s"
-        % (
-            messages.TOKEN_HAS_EXPIRED,
-            messages.TOKEN_IS_INVALID,
-            messages.AUTHORISATION_TOKEN_IS_MISSING,
-            messages.USER_NOT_INVOLVED_IN_THIS_MENTOR_RELATION,
-        ),
-    )
-    @mentorship_relation_ns.response(
-        HTTPStatus.NOT_FOUND,
-        "%s\n%s"
-        % (messages.MENTORSHIP_RELATION_DOES_NOT_EXIST, messages.TASK_DOES_NOT_EXIST),
-    )
-    def delete(cls, request_id, task_id):
-        """
-        Delete a task.
-
-        Input:
-        1. Header: valid access token
-        2. Path: ID of the task to be deleted (task_id) and it ID of the associated
-        mentorship relation (request_id).
-        3. Body: JSON object containing description of task.
-
-        Returns:
-        Success or failure message. Task is deleted if request is successful.
-        """
-
-        # TODO check if user id is well parsed, if it is an integer
-
-        user_id = get_jwt_identity()
-
-        response = TaskDAO.delete_task(
-            user_id=user_id, mentorship_relation_id=request_id, task_id=task_id
-        )
-
-        return response
-
-
-@mentorship_relation_ns.route("mentorship_relation/<int:request_id>/tasks")
-class ListTasks(Resource):
-    @classmethod
-    @jwt_required
-    @mentorship_relation_ns.doc("list_tasks_in_mentorship_relation")
-    @mentorship_relation_ns.expect(auth_header_parser)
-    @mentorship_relation_ns.response(
-        HTTPStatus.OK,
-        "List tasks from a mentorship relation with success.",
-        model=list_tasks_response_body,
-    )
-    @mentorship_relation_ns.response(
-        HTTPStatus.UNAUTHORIZED,
-        "%s\n%s\n%s\n%s"
-        % (
-            messages.TOKEN_HAS_EXPIRED,
-            messages.TOKEN_IS_INVALID,
-            messages.AUTHORISATION_TOKEN_IS_MISSING,
-            messages.USER_NOT_INVOLVED_IN_THIS_MENTOR_RELATION,
-        ),
-    )
-    @mentorship_relation_ns.response(
-        HTTPStatus.NOT_FOUND, "%s" % messages.MENTORSHIP_RELATION_DOES_NOT_EXIST
-    )
-    def get(cls, request_id):
-        """
-        List all tasks from a mentorship relation.
-
-        Input:
-        1. Header: valid access token
-        2. Path: ID of the mentorship relation for which tasks are to be
-        displayed(request_id). The user must be involved in this relation.
-
-        Returns:
-        JSON array containing task details as objects is displayed on success.
-        """
-
-        # TODO check if user id is well parsed, if it is an integer
-
-        user_id = get_jwt_identity()
-
-        response = TaskDAO.list_tasks(
-            user_id=user_id, mentorship_relation_id=request_id
-        )
-
-        if isinstance(response, tuple):
-            return response
-
-        return marshal(response, list_tasks_response_body), HTTPStatus.OK
-
-
-@mentorship_relation_ns.route(
-    "mentorship_relation/<int:request_id>/task/<int:task_id>/complete"
-)
-class UpdateTask(Resource):
-    @classmethod
-    @jwt_required
-    @mentorship_relation_ns.doc("update_task_in_mentorship_relation")
-    @mentorship_relation_ns.expect(auth_header_parser)
-    @mentorship_relation_ns.response(
-        HTTPStatus.OK, "%s" % messages.TASK_WAS_ACHIEVED_SUCCESSFULLY
-    )
-    @mentorship_relation_ns.response(HTTPStatus.BAD_REQUEST, "%s" % messages.TASK_WAS_ALREADY_ACHIEVED)
-    @mentorship_relation_ns.response(
-        HTTPStatus.UNAUTHORIZED,
-        "%s\n%s\n%s\n%s"
-        % (
-            messages.TOKEN_HAS_EXPIRED,
-            messages.TOKEN_IS_INVALID,
-            messages.AUTHORISATION_TOKEN_IS_MISSING,
-            messages.USER_NOT_INVOLVED_IN_THIS_MENTOR_RELATION,
-        ),
-    )
-    @mentorship_relation_ns.response(
-        HTTPStatus.NOT_FOUND,
-        "%s\n%s"
-        % (messages.MENTORSHIP_RELATION_DOES_NOT_EXIST, messages.TASK_DOES_NOT_EXIST),
-    )
-    def put(cls, request_id, task_id):
-        """
-        Update a task to mark it as complate
-
-        Input:
-        1. Header: valid access token
-        2. Path: ID of task (task_id) and ID of the associated mentorship
-        relation (request_id). The user must be involved in this relation.
-        3. Body:
-
-        Returns:
-        Success or failure message. The task is marked as complete if succesful.
-        """
-
-        # TODO check if user id is well parsed, if it is an integer
-
-        user_id = get_jwt_identity()
-
-        response = TaskDAO.complete_task(
-            user_id=user_id, mentorship_relation_id=request_id, task_id=task_id
-        )
-
-        return response
-
-
-@mentorship_relation_ns.route(
-    "mentorship_relation/<int:relation_id>/task/<int:task_id>/comment"
-)
-class CreateTaskComment(Resource):
-    @classmethod
-    @jwt_required
-    @mentorship_relation_ns.expect(auth_header_parser, task_comment_model)
-    @mentorship_relation_ns.doc(
-        responses={
-            HTTPStatus.CREATED:f"{messages.TASK_COMMENT_WAS_CREATED_SUCCESSFULLY}",
-            HTTPStatus.BAD_REQUEST: f"{messages.COMMENT_FIELD_IS_MISSING}<br>"
-            f"{messages.COMMENT_NOT_IN_STRING_FORMAT}<br>"
-            f"{ {'message': get_length_validation_error_message('comment', None, COMMENT_MAX_LENGTH)}}<br>"
-            f"{messages.UNACCEPTED_STATE_RELATION}",
-            HTTPStatus.UNAUTHORIZED: f"{messages.TOKEN_HAS_EXPIRED}<br>"
-            f"{messages.TOKEN_IS_INVALID}<br>"
-            f"{messages.AUTHORISATION_TOKEN_IS_MISSING}<br>"
-            f"{messages.USER_NOT_INVOLVED_IN_THIS_MENTOR_RELATION}",
-            HTTPStatus.NOT_FOUND: f"{messages.USER_DOES_NOT_EXIST}<br>"
-            f"{messages.MENTORSHIP_RELATION_DOES_NOT_EXIST}<br>"
-            f"{messages.TASK_DOES_NOT_EXIST}",
-        }
-    )
-    def post(cls, relation_id, task_id):
-        """
-        Creates a new task comment.
-        """
-
-        data = request.json
-
-        is_valid = validate_task_comment_request_data(data)
-        if is_valid != {}:
-            return is_valid, HTTPStatus.BAD_REQUEST
-
-        comment = data["comment"]
-        return TaskCommentDAO.create_task_comment(
-            get_jwt_identity(), task_id, relation_id, comment
-        )
-
-
-@mentorship_relation_ns.route(
-    "mentorship_relation/<int:relation_id>/task/<int:task_id>/comment/<int:comment_id>"
-)
-class TaskComment(Resource):
-    @classmethod
-    @jwt_required
-    @mentorship_relation_ns.expect(auth_header_parser, task_comment_model)
-    @mentorship_relation_ns.doc(
-        responses={
-            HTTPStatus.OK: f"{messages.TASK_COMMENT_WAS_UPDATED_SUCCESSFULLY}",
-            HTTPStatus.BAD_REQUEST: f"{messages.COMMENT_FIELD_IS_MISSING}<br>"
-            f"{messages.COMMENT_NOT_IN_STRING_FORMAT}<br>"
-            f"{ {'message':get_length_validation_error_message('comment', None, COMMENT_MAX_LENGTH)}}<br>"
-            f"{messages.UNACCEPTED_STATE_RELATION}<br>"
-            f"{messages.TASK_COMMENT_WAS_NOT_CREATED_BY_YOU}",
-            HTTPStatus.UNAUTHORIZED: f"{messages.TOKEN_HAS_EXPIRED}<br>"
-            f"{messages.TOKEN_IS_INVALID}<br>"
-            f"{messages.AUTHORISATION_TOKEN_IS_MISSING}<br>"
-            f"{messages.USER_NOT_INVOLVED_IN_THIS_MENTOR_RELATION}",
-            HTTPStatus.NOT_FOUND: f"{messages.USER_DOES_NOT_EXIST}<br>"
-            f"{messages.MENTORSHIP_RELATION_DOES_NOT_EXIST}<br>"
-            f"{messages.TASK_DOES_NOT_EXIST}<br>"
-            f"{messages.TASK_COMMENT_DOES_NOT_EXIST}<br>"
-            f"{messages.TASK_COMMENT_WITH_GIVEN_TASK_ID_DOES_NOT_EXIST}",
-        }
-    )
-    def put(cls, relation_id, task_id, comment_id):
-        """
-        Modifies the task comment.
-        """
-
-        data = request.json
-
-        is_valid = validate_task_comment_request_data(data)
-        if is_valid != {}:
-            return is_valid, HTTPStatus.BAD_REQUEST
-
-        comment = data["comment"]
-        return TaskCommentDAO.modify_comment(
-            get_jwt_identity(), comment_id, task_id, relation_id, comment
-        )
-
-    @classmethod
-    @jwt_required
-    @mentorship_relation_ns.expect(auth_header_parser)
-    @mentorship_relation_ns.doc(
-        responses={
-            HTTPStatus.OK: f"{messages.TASK_COMMENT_WAS_DELETED_SUCCESSFULLY}",
-            HTTPStatus.BAD_REQUEST: f"{messages.UNACCEPTED_STATE_RELATION}",
-            HTTPStatus.UNAUTHORIZED: f"{messages.TOKEN_HAS_EXPIRED}<br>"
-            f"{messages.TOKEN_IS_INVALID}<br>"
-            f"{messages.AUTHORISATION_TOKEN_IS_MISSING}<br>"
-            f"{messages.USER_NOT_INVOLVED_IN_THIS_MENTOR_RELATION}",
-            HTTPStatus.FORBIDDEN: f"{messages.TASK_COMMENT_WAS_NOT_CREATED_BY_YOU_DELETE}",
-            HTTPStatus.NOT_FOUND: f"{messages.USER_DOES_NOT_EXIST}<br>"
-            f"{messages.MENTORSHIP_RELATION_DOES_NOT_EXIST}<br>"
-            f"{messages.TASK_DOES_NOT_EXIST}<br>"
-            f"{messages.TASK_COMMENT_DOES_NOT_EXIST}<br>"
-            f"{messages.TASK_COMMENT_WITH_GIVEN_TASK_ID_DOES_NOT_EXIST}",
-        }
-    )
-    def delete(cls, relation_id, task_id, comment_id):
-        """
-        Deletes the task comment.
-        """
-
-        return TaskCommentDAO.delete_comment(
-            get_jwt_identity(), comment_id, task_id, relation_id
-        )
-
-
-@mentorship_relation_ns.route(
-    "mentorship_relation/<int:relation_id>/task/<int:task_id>/comments/"
-)
-class TaskComments(Resource):
-    @classmethod
-    @jwt_required
-    @mentorship_relation_ns.expect(auth_header_parser)
-    @mentorship_relation_ns.response(
-        HTTPStatus.OK, f"{ messages.LIST_TASK_COMMENTS_WITH_SUCCESS}", task_comments_model
-    )
-    @mentorship_relation_ns.doc(
-        responses={
-            HTTPStatus.BAD_REQUEST: f"{messages.UNACCEPTED_STATE_RELATION}",
-            HTTPStatus.UNAUTHORIZED: f"{messages.TOKEN_HAS_EXPIRED}<br>"
-            f"{messages.TOKEN_IS_INVALID}<br>"
-            f"{messages.AUTHORISATION_TOKEN_IS_MISSING}<br>"
-            f"{messages.USER_NOT_INVOLVED_IN_THIS_MENTOR_RELATION}",
-            HTTPStatus.NOT_FOUND: f"{messages.USER_DOES_NOT_EXIST}<br>"
-            f"{messages.MENTORSHIP_RELATION_DOES_NOT_EXIST}<br>"
-            f"{messages.TASK_DOES_NOT_EXIST}",
-        }
-    )
-    def get(cls, relation_id, task_id):
-        """
-        Lists the task comments.
-        """
-
-        response = TaskCommentDAO.get_all_task_comments_by_task_id(
-            get_jwt_identity(), task_id, relation_id
-        )
-
-        if isinstance(response, tuple):
-            return response
-        else:
-            return marshal(response, task_comments_model), HTTPStatus.OK
